+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - use varint when necessary
+- add empty extensions encoders for some operation types
 
 ## [1.1.0] - 2022-04-13